# backend-api/docs/swagger.yaml
swagger: '2.0'
info:
    title: RecipeBlog
    version: 1.0.0
    description: This is a blog for recipes
paths:
...
    '/recipes/{RecipeID}':
        get:
            tags:
                - recipes
            parameters:
                -
                    name: RecipeID
                    description: ID of <recipe>
                    in: path
                    required: true
                    type: string
            responses:
                '200':
                    description: Returns all details of the given recipe.
                    schema:
                        type: array
                        items:
                            type: string
                    examples:
                        application/json:
                            -
                                RecipeID: 5fa725fd-cec3-11f0-b43f-901b0eec5991
                                Title: TestRecipe
                                Content: For testing
                                createdAt: '2025-12-01T16:37:08.000Z'
                                updatedAt: '2025-12-01T16:37:08.000Z'
                '404':
                    description: Returns error when recipe not found.
                    schema:
                        $ref: '#/definitions/Error'
            operationId: RecipeID
            summary: Get full detail of a recipe
            description: Will return all details of a given recipe based on RecipeID given.
        put:
            tags:
                - recipes
            parameters:
                -
                    name: body
                    description: |
                        Contains new data to be changed in the specified Recipe
                    schema:
                        $ref: '#/definitions/New-Recipe'
                    in: body
                    required: true
            responses:
                '201':
                    description: Returns 201 upon successsful modification of recipe
                    schema:
                        $ref: '#/definitions/New-Recipe'
                    examples:
                        application/json:
                            -
                                Title: '0'
                                Content: abc
                                Image: im2gf
                '400':
                    description: Returns 400 when one or multile required parameters are missing
                    schema:
                        $ref: '#/definitions/Error'
                    examples:
                        application/json:
                            -
                                error: 'cannot change info, because some parameters are missing'
                '404':
                    description: Returns 404 when Recipe to be modified is not found by the given RecipeID
                    schema:
                        $ref: '#/definitions/Error'
                    examples:
                        application/json:
                            -
                                error: Recipe not found
        delete:
            tags:
                - recipes
            responses:
                '204':
                    description: Returns "No Content" when removal of recipes is successful
                    schema:
                        $ref: '#/definitions/Error'
                    examples:
                        application/json:
                            -
                                error: No countent
                '404':
                    description: 'Returns error, when recipe with specified ID does not exist'
                    schema:
                        $ref: '#/definitions/Error'
                    examples:
                        application/json:
                            -
                                error: Not found
            summary: Delete a recipe from the app
            description: |
                Deletes an entry from the app
        parameters:
            -
                name: RecipeID
                description: |
                    ID of recipe
                in: path
                required: true
                type: string
...
definitions:
    Error:
        title: Root Type for Error
        description: ''
        type: object
        properties:
            error:
                type: string
        example:
            error: Error message
    New-Recipe:
        title: Root Type for New-Recipe
        description: ''
        required:
            - Content
            - Title
        type: object
        properties:
            Title:
                type: string
            Content:
                type: string
            Image:
                type: string
        example:
            Title: NewRecipeTitle
            Content: NewRecipeContent
            Image: NewRecipeImage
...